+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/audiovideo/video/DefaultVideoTileFactory.kt
@@ -8,7 +8,12 @@ package com.amazonaws.services.chime.sdk.meetings.audiovideo.video
 import com.amazonaws.services.chime.sdk.meetings.utils.logger.Logger
 
 class DefaultVideoTileFactory(private val logger: Logger) : VideoTileFactory {
-    override fun makeTile(tileId: Int, attendeeId: String?): VideoTile {
-        return DefaultVideoTile(logger, tileId, attendeeId)
+    override fun makeTile(
+        tileId: Int,
+        attendeeId: String?,
+        videoStreamContentHeight: Int,
+        videoStreamContentWidth: Int
+    ): VideoTile {
+        return DefaultVideoTile(logger, tileId, attendeeId, videoStreamContentHeight, videoStreamContentWidth)
     }
 }