+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/audiovideo/video/VideoTileController.kt
@@ -24,14 +24,16 @@ interface VideoTileController : VideoTileControllerFacade {
      * Called whenever there is a new Video frame received for any of the attendee in the meeting
      *
      * @param frame: [Any] - A frame of video
-     * @param attendeeId: [String] - An id of user who is transmitting current frame
-     * @param pauseState: [VideoPauseState] - Current pause state of the video being received
      * @param videoId: [Int] - Unique id that belongs to video being transmitted
+     * @param attendeeId: [String] - An id of user who is transmitting current frame
+     * @param videoStreamContentHeight - Height of the video stream being transmitted
+     * @param videoStreamContentWidth - Width of the video stream being transmitted
+     * @param pauseState: [VideoPauseState] - Current pause state of the video being received
      */
     fun onReceiveFrame(
         frame: Any?,
+        videoId: Int,
         attendeeId: String?,
-        pauseState: VideoPauseState,
-        videoId: Int
+        pauseState: VideoPauseState
     )
 }