+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/audiovideo/video/VideoTileFactory.kt
@@ -14,5 +14,10 @@ interface VideoTileFactory {
      *
      * @return [VideoTile] to use with [VideoTileController]
     */
-    fun makeTile(tileId: Int, attendeeId: String?): VideoTile
+    fun makeTile(
+        tileId: Int,
+        attendeeId: String?,
+        videoStreamContentHeight: Int,
+        videoStreamContentWidth: Int
+    ): VideoTile
 }