+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/audiovideo/video/VideoTileState.kt
@@ -10,10 +10,12 @@ package com.amazonaws.services.chime.sdk.meetings.audiovideo.video
  *
  * @property tileId: Int - Unique Id associated with this tile
  * @property attendeeId: String? - Attendee Id of the user associated with this tile
+ * @property videoStreamContentHeight: Int - Height of video stream content
+ * @property videoStreamContentWidth: Int - Width of video stream content
  * @property pauseState: VideoPauseState - The current pause state of the tile
  * @property isLocalTile: Boolean - Whether the video tile is for the local attendee
  */
-data class VideoTileState(val tileId: Int, val attendeeId: String?, var pauseState: VideoPauseState) {
+data class VideoTileState(val tileId: Int, val attendeeId: String?, var videoStreamContentHeight: Int, var videoStreamContentWidth: Int, var pauseState: VideoPauseState) {
     /**
      * Whether the video tile is for the local attendee
      */