+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 * Added choose device list in the demo app for speaker button
 * Added metrics in the demo app
+* Added `onVideoTileSizeChanged` API in `VideoTileObserver` for video stream content size change
 
 ### Changed
 * Changed to layout buttons instead of tabs